+++ examples/react/src/main.tsx
@@ -2,6 +2,8 @@ import React from 'react';
 import { Welcome } from './components/index';
 import './main.css';
 
+declare const BTN: string;
+declare const IRRELEVANT_ESCAPE_ENV: string;
 export function Main() {
   return (
     <>