+++ examples/solid/farm.config.ts
@@ -1,7 +1,14 @@
+import type { UserConfig } from '@farmfe/core';
 import farmJsPluginSolid from '@farmfe/js-plugin-solid';
 
-export default {
+function defineFarmConfig(config: UserConfig) {
+  return config;
+}
+
+export default defineFarmConfig({
   compilation: {
+    minify: false,
+    presetEnv: false,
     mode: process.env.NODE_ENV,
     input: {
       index: 'index.html'
@@ -13,5 +20,8 @@ export default {
       __DEV__: 'true'
     }
   },
+  server: {
+    open: true
+  },
   plugins: [farmJsPluginSolid()]
-};
+});