+++ js-plugins/solid/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @farmfe/js-plugin-solid
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [56f235c]
+  - @farmfe/core@0.11.0
+
+## 1.0.3
+
+### Patch Changes
+
+- 4e14bc3: fix: hmr assertion
+
 ## 1.0.2
 
 ### Patch Changes