+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @farmfe/core
 
+## 0.11.0
+
+### Minor Changes
+
+- 56f235c: Upgrade swc crates and support emotion
+
+## 0.10.7
+
+### Patch Changes
+
+- 75f58c1: Fix that extra watch file panic
+
 ## 0.10.6
 
 ### Patch Changes