+++ rust-plugins/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @farmfe/plugin-react
 
+## 0.3.0
+
+### Minor Changes
+
+- 56f235c: Upgrade swc crates and support emotion
+
+## 0.2.1
+
+### Patch Changes
+
+- 75f58c1: Fix that extra watch file panic
+
 ## 0.2.0
 
 ### Minor Changes